+++ src/chartreuse/tests/e2e_tests/sample_alembic/versions/init.py
@@ -1,0 +1,27 @@
+"""init
+
+Revision ID: 78db1f820ded
+Revises:
+Create Date: 2019-02-14 10:14:38.289690
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = "aaaaaaaaaaaa"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    from alembic import op
+
+    op.create_table(
+        "migrated",
+    )
+
+
+def downgrade():
+    from alembic import op
+
+    op.drop_table("migrated")