+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-# 1.0.1 (2020-04-10)
+# 2.0.0 (Unreleased)
+## BREAKING CHANGES
+- helm: move most alembic/eslembic parameters to eslembic and alembic prefix.
+## Features
+- eslembic: Split eslembic upgrade to eslembic migrate and eslembic upgrade as a post job
+- eslembic: allow to disable/enable eslembic upgrade
 ## Fixes
-- setup.py: prevent upgrade to incompatible major version of wiremind-kube.
+- Upgrade to eslembic v6.x.x
+
+# 1.1.0 (2020-03-10)
+## Features
+## Fixes:
+- Set terminationGracePeriodSeconds: 0 in all jobs.
 
 # 1.0.0 (2020-02-03)
 Please note that python code and helm chart versions are now synchronized.